+++ raise_utils/metrics/metrics.py
@@ -6,6 +6,7 @@ from raise_utils.metrics.impl import get_roc_auc
 from raise_utils.metrics.impl import get_d2h, get_d2h2
 from raise_utils.metrics.impl import get_f1_score
 from raise_utils.metrics.impl import get_precision
+from raise_utils.metrics.impl import get_pd_pf
 from raise_utils.metrics.impl import get_confusion_matrix
 import numpy as np
 
@@ -21,6 +22,7 @@ name_map = {
     "d2h": get_d2h,
     "f1": get_f1_score,
     "prec": get_precision,
+    "pd-pf": get_pd_pf,
     "conf": get_confusion_matrix
 }